requirements.txt
aiohttp @ file:///C:/b/abs_c4zmy2l696/croot/aiohttp_1670009573673/work
aiosignal @ file:///tmp/build/80754af9/aiosignal_1637843061372/work
alembic==1.11.1
async-timeout @ file:///C:/b/abs_43ozhz2a8g/croots/recipe/async-timeout_1664876362767/work
attrs @ file:///C:/b/abs_09s3y775ra/croot/attrs_1668696195628/work
beautifulsoup4==4.12.2
blinker==1.6.2
brotlipy==0.7.0
bs4==0.0.1
certifi @ file:///C:/b/abs_36eb5mzhph/croot/certifi_1690232276943/work/certifi
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.6
colorama==0.4.6
cryptography @ file:///C:/b/abs_13590mi9q9/croot/cryptography_1689373706078/work
d7==1.0.1
dnspython==2.4.1
elasticsearch @ file:///tmp/build/80754af9/elasticsearch_1625769069476/work
et-xmlfile==1.1.0
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
frozenlist @ file:///C:/b/abs_2bb5uzghsi/croot/frozenlist_1670004511812/work
greenlet==2.0.2
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
itsdangerous==2.1.2
Jinja2 @ file:///C:/b/abs_7cdis66kl9/croot/jinja2_1666908141852/work
joblib==1.3.2
Mako==1.2.4
MarkupSafe @ file:///C:/ci/markupsafe_1654508036328/work
mongoengine==0.27.0
multidict @ file:///C:/b/abs_6cx_8w3cv2/croot/multidict_1665674238352/work
nltk==3.8.1
numpy==1.26.0
openpyxl==3.0.10
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyJWT @ file:///C:/ci/pyjwt_1657529477795/work
pymongo==3.12.0
pyOpenSSL @ file:///C:/b/abs_08f38zyck4/croot/pyopenssl_1690225407403/work
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
regex==2023.8.8
requests @ file:///C:/b/abs_316c2inijk/croot/requests_1690400295842/work
soupsieve==2.5
SQLAlchemy==2.0.19
tqdm==4.66.1
typing_extensions==4.7.1
urllib3 @ file:///C:/b/abs_889_loyqv4/croot/urllib3_1686163174463/work
Werkzeug==2.3.6
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
WTForms==2.3.3
xlrd @ file:///croot/xlrd_1685030938141/work
yarl @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_e5nlunspj6/croots/recipe/yarl_1661437086516/work
